+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,12 +2,18 @@
 
 ## Unreleased
 
+* Added
+  * Migrate doc from terra-dropdown-button.
+
 ## 1.12.0 - (November 2, 2021)
 
 * Added
   * Migrate docs from terra-cell-grid.
   * Migrate docs from terra-demographics-banner.
   * Migrate docs from terra-content-container.
+  * Migrate doc from terra-card.
+  * Migrate doc from terra-button.
+  * Migrate doc from terra-button-group.
   * Migrate docs from terra-card.
   * Migrate docs from terra-button.
   * Migrate docs from terra-button-group.